feat!: merge analyze_engagement/network/suggest_content_strategy into analyze_account
